remove xr.differentiate
