+++ doc/api.rst
@@ -25,7 +25,6 @@ Top-level functions
    zeros_like
    ones_like
    dot
-   differentiate
 
 Dataset
 =======
